+++ src/main/java/dopamine/backend/backoffice/controller/FeedBackofficeController.java
@@ -1,0 +1,20 @@
+package dopamine.backend.backoffice.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+
+@Controller
+@EnableWebMvc
+@RequestMapping("/backoffice/feed")
+@RequiredArgsConstructor
+public class FeedBackofficeController {
+    @GetMapping
+    public String feed(Model model) {
+        model.addAttribute("data", "안녕");
+        return "feed/feed";
+    }
+}
